walt_restauth_1/core/requirements.txt
asgiref==3.6.0
attrs==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
dj-rest-auth==4.0.0
Django==4.2.1
django-allauth==0.52.0
django-extensions==3.2.1
django-phonenumber-field==7.1.0
django-rest-authtoken==2.1.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.4
Faker==18.13.0
httpie==3.2.2
idna==3.4
inflection==0.5.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
markdown-it-py==2.2.0
mdurl==0.1.2
multidict==6.0.4
oauthlib==3.2.2
phonenumbers==8.13.11
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycparser==2.21
Pygments==2.15.1
PyJWT==2.6.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.0
requests==2.30.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rest-framework-simplejwt==0.0.2
rich==13.4.1
rpds-py==0.9.2
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.0.2
